+++ WeatherApp/Common/Sources/Views/Degree.swift
@@ -1,0 +1,10 @@
+//
+//  Degree.swift
+//  WeatherApp
+//
+//  Created by AndUser on 06.04.2021.
+//
+
+enum Degree: String {
+    case sign = "°"
+}
