+++ WeatherApp/Modules/MainScreen/View/Components/Hourly View/HourlyView.swift
@@ -15,7 +15,6 @@ final class HourlyView: UICollectionView {
 
     override init(frame: CGRect, collectionViewLayout layout: UICollectionViewLayout) {
         super.init(frame: frame, collectionViewLayout: layout)
-        print(cellId)
 
         setupView()
     }
